TextureReplacement: Import texture array created from editor
Imports texture arrays created from the Unity Editor as an alternative to make one at runtime from individual textures.
This is faster, does not require to load and discharge original textures and avoid problems that causes the creation process to fail (size/format mismatch between textures).

The downside is that Unity provides this feature only from its API, but there are free tools on the Asset Store that adds an inspector UI.